Adding gcc 2.95.2

+++ entrypoint.sh
@@ -1,5 +1,11 @@
 #!/bin/sh
 
+if [ ! -e "/build" ]; then
+    echo "Mount /build when starting container!"
+    echo "  e.g. docker run --rm -v \$(pwd):/build old-gcc"
+    exit 0
+fi
+
 cd /work/gcc*
 mv xgcc gcc
 cp cpp cc1 gcc cc1plus g++ /build/
